refactor(station): group Host and PTY tests
